Extract address credential check to shared helper
